Use single fuzzer context
