+++ src/main/kotlin/kotlinx/fuzzer/Fuzzer.kt
@@ -1,10 +1,10 @@
 package kotlinx.fuzzer
 
+import kotlinx.fuzzer.fuzzing.inputhandlers.FuzzerContext
 import kotlinx.fuzzer.fuzzing.inputhandlers.InputTask
 import kotlinx.fuzzer.fuzzing.inputhandlers.MutationTask
 import kotlinx.fuzzer.fuzzing.log.Logger
 import kotlinx.fuzzer.fuzzing.log.TasksLog
-import kotlinx.fuzzer.fuzzing.storage.ContextFactory
 import kotlinx.fuzzer.fuzzing.storage.Storage
 import kotlinx.fuzzer.fuzzing.storage.createStorageStrategy
 import java.io.File
@@ -24,8 +24,8 @@ class Fuzzer(arguments: FuzzerArgs) {
         Logger(storage, stop, File(arguments.workingDirectory), log)
     }
     private val storage = Storage(this, File(arguments.workingDirectory), arguments.storageStrategy)
-    internal val contextFactory = ContextFactory(this, storage, arguments)
-    private val mutationTask = MutationTask(this, storage, contextFactory)
+    internal val context = FuzzerContext(storage, arguments, this)
+    private val mutationTask = MutationTask(this, storage, context)
     private val stop = AtomicBoolean(false)
     private var exception: Throwable? = null
 
@@ -33,7 +33,7 @@ class Fuzzer(arguments: FuzzerArgs) {
 
     fun start(timeout: Long? = null, unit: TimeUnit = TimeUnit.SECONDS) {
         mutationTask.start()
-        storage.listCorpusInput().map { InputTask(contextFactory, it) }.forEach { submit(it) }
+        storage.listCorpusInput().map { InputTask(context, it) }.forEach { submit(it) }
         submit(Runnable { logger.log("All init corpus submitted") })
         setUpTimeTimeout(timeout, unit)
         runCatching { logger.run() }.onFailure { e -> stop(e) }
