+++ src/main/kotlin/kotlinx/fuzzer/fuzzing/inputhandlers/FuzzerContext.kt
@@ -1,22 +1,20 @@
 package kotlinx.fuzzer.fuzzing.inputhandlers
 
-import kotlinx.fuzzer.coverage.createCoverageRunner
 import kotlinx.fuzzer.Fuzzer
 import kotlinx.fuzzer.FuzzerArgs
+import kotlinx.fuzzer.coverage.createCoverageRunner
 import kotlinx.fuzzer.fuzzing.TargetMethod
-import kotlinx.fuzzer.fuzzing.storage.ContextFactory
 import kotlinx.fuzzer.fuzzing.storage.Storage
 
-/** Set of services for independent worker execution. */
+/** Set of services for worker execution. */
 class FuzzerContext(
     val storage: Storage,
     arguments: FuzzerArgs,
-    fuzzer: Fuzzer,
-    contextFactory: ContextFactory
+    fuzzer: Fuzzer
 ) {
     val coverageRunner = createCoverageRunner(arguments.classpath, arguments.packages)
     val targetMethod: TargetMethod
-    val mutator = InputMutator(fuzzer, storage, contextFactory, 1)
+    val mutator = InputMutator(fuzzer, storage, this, 1)
     val compositeCoverageCount = arguments.compositeCoverageCount
 
     init {
