+++ src/main/kotlin/kotlinx/fuzzer/fuzzing/mutation/MutationFactory.kt
@@ -7,8 +7,8 @@ import kotlinx.fuzzer.fuzzing.storage.Storage
  * All mutations container.
  * @see  <a href="https://github.com/dvyukov/go-fuzz">go-fuzz</a>
  */
-class MutationFactory(storage: Storage) : Mutation {
-    private val mutations = listOf(
+class MutationFactory(private val storage: Storage) : Mutation {
+    private fun mutations() = listOf(
         InsertBytesMutation(),
         InsertCharsMutation(),
         RemoveBytesMutation(),
@@ -31,6 +31,11 @@ class MutationFactory(storage: Storage) : Mutation {
         SpliceAnotherInputMutation(storage),
         ReplaceTextNumberMutation()
     )
+
+    /** Some mutations have state(@see [ReplaceNumberMutation]), so their usage is not thread-safe. */
+    private val threadLocalMutations = ThreadLocal.withInitial { mutations() }
+    private val mutations: List<Mutation>
+        get() = threadLocalMutations.get()
 
     override fun mutate(bytes: ByteArray) = mutations.random().mutate(bytes)
