+++ src/test/kotlin/kotlinx/fuzzer/fuzzing/timeouttest/FuzzerTimeoutTest.kt
@@ -1,0 +1,26 @@
+package kotlinx.fuzzer.fuzzing.timeouttest
+
+import kotlinx.fuzzer.Fuzz
+import kotlinx.fuzzer.FuzzCrash
+import kotlinx.fuzzer.Fuzzer
+import org.junit.jupiter.api.Assertions
+import org.junit.jupiter.api.Test
+import java.io.File
+
+internal class FuzzerTimeoutTest {
+    @Fuzz("testB")
+    fun fuzz(bytes: ByteArray): Int {
+        return 1
+    }
+
+    @FuzzCrash
+    fun callback(e: Throwable, data: ByteArray) {
+        throw e
+    }
+
+    @Test
+    fun timeoutTest() {
+        Fuzzer(FuzzerTimeoutTest::class.java).start(2)
+        Assertions.assertTrue(File("testB").deleteRecursively())
+    }
+}
